requirements.txt
# Core
python-dotenv
streamlit
tqdm

# LangChain + Embeddings + Vector Store
langchain
langchain-community
langchain-core
faiss-cpu
sentence-transformers

# LLMs & Chat Models
openai
google-generativeai

# Multi-Agent Framework
langgraph

# Web Search Tool
duckduckgo-search

# PDF and Document Loaders
PyPDF2
docx2txt
markdown
python-docx

# Utility
requests
numpy
pandas

unstructured